fix: change 'AnimalNotFountException' to 'AnimalNotFoundException', change all instance where the 'price' is 'int' to 'double',use type of animal's Class to search for animal,use 'instance' to determine if they are the same

+++ work1/Wea-sudo/AnimalShop/Bill.java
@@ -5,10 +5,10 @@ import java.time.LocalDate;
 public class Bill {
     private Customer customer;
     private Animal animal;
-    private int price;
+    private double price;
     private LocalDate localDate;
 
-    public Bill(Customer customer, Animal animal, int price, LocalDate localDate) {
+    public Bill(Customer customer, Animal animal, double price, LocalDate localDate) {
         this.customer = customer;
         this.animal = animal;
         this.price = price;
